requirements.txt
streamlit==1.32.*
requests>=2.27,<3
python-dotenv>=1.0
pandas>=1.3,<3
nltk>=3.8
